made global func for password change
